AG-17010 Move link drawing to NetworkTreeLayout
