+++ packages/ag-charts-enterprise/src/series/network/networkTreeLayout.ts
@@ -1,5 +1,5 @@
 import { _ModuleSupport } from 'ag-charts-community';
-import type { Vertex } from 'ag-charts-core';
+import { Vec2, type Vertex } from 'ag-charts-core';
 
 import type { NetworkGraph } from './networkGraph';
 import { NetworkLayout } from './networkLayout';
@@ -25,7 +25,10 @@ export class NetworkTreeLayout<TVertex, TEdge> extends NetworkLayout<TVertex, TE
         vertices: Vertex<TVertex, TEdge>[],
         getDatumNodeBBox: (vertex: Vertex<TVertex, TEdge>) => TBBox | undefined,
         layoutDatumNode: (vertex: Vertex<TVertex, TEdge>, groupBBox: TBBox) => void,
-        layoutLinkNode: (vertex: Vertex<TVertex, TEdge>, parentBBox: TBBox, childBBox: TBBox) => void
+        layoutLinkNode: (
+            vertex: Vertex<TVertex, TEdge>,
+            drawLink: (path: _ModuleSupport.ExtendedPath2D) => void
+        ) => void
     ) {
         this.updateChildren(graph, vertices, getDatumNodeBBox, layoutDatumNode, layoutLinkNode);
 
@@ -37,7 +40,10 @@ export class NetworkTreeLayout<TVertex, TEdge> extends NetworkLayout<TVertex, TE
         vertices: Vertex<TVertex, TEdge>[],
         getDatumNodeBBox: (vertex: Vertex<TVertex, TEdge>) => TBBox | undefined,
         layoutDatumNode: (vertex: Vertex<TVertex, TEdge>, groupBBox: TBBox) => void,
-        layoutLinkNode: (vertex: Vertex<TVertex, TEdge>, parentBBox: TBBox, childBBox: TBBox) => void,
+        layoutLinkNode: (
+            vertex: Vertex<TVertex, TEdge>,
+            drawLink: (path: _ModuleSupport.ExtendedPath2D) => void
+        ) => void,
         groupBBox: TBBox = new BBox(0, 0, 0, 0)
     ): {
         containerBBox: TBBox;
@@ -97,7 +103,9 @@ export class NetworkTreeLayout<TVertex, TEdge> extends NetworkLayout<TVertex, TE
             // Request the series to layout the links between children and their parents.
             if (childrenBBoxes) {
                 for (const { vertex: childVertex, bbox } of childrenBBoxes) {
-                    layoutLinkNode(childVertex, layoutBBox, bbox);
+                    layoutLinkNode(childVertex, (path: _ModuleSupport.ExtendedPath2D) =>
+                        this.drawLink(path, layoutBBox, bbox)
+                    );
                 }
             }
         }
@@ -113,7 +121,10 @@ export class NetworkTreeLayout<TVertex, TEdge> extends NetworkLayout<TVertex, TE
         vertex: Vertex<TVertex, TEdge>,
         getDatumNodeBBox: (vertex: Vertex<TVertex, TEdge>) => TBBox | undefined,
         layoutDatumNode: (vertex: Vertex<TVertex, TEdge>, groupBBox: TBBox) => void,
-        layoutLinkNode: (vertex: Vertex<TVertex, TEdge>, parentBBox: TBBox, childBBox: TBBox) => void,
+        layoutLinkNode: (
+            vertex: Vertex<TVertex, TEdge>,
+            drawLink: (path: _ModuleSupport.ExtendedPath2D) => void
+        ) => void,
         groupBBox: TBBox,
         datumBBox?: TBBox
     ) {
@@ -137,5 +148,19 @@ export class NetworkTreeLayout<TVertex, TEdge> extends NetworkLayout<TVertex, TE
             childrenBBoxes,
             mergedChildrenBBoxes: BBox.merge(childrenBBoxes.map(({ bbox }) => bbox)),
         };
+    }
+
+    private drawLink(path: _ModuleSupport.ExtendedPath2D, parentBBox: TBBox, childBBox: TBBox) {
+        const start = Vec2.from(parentBBox.x + parentBBox.width / 2, parentBBox.y + parentBBox.height);
+        const end = Vec2.from(childBBox.x + childBBox.width / 2, childBBox.y);
+        const elbowDist = Vec2.from(0, (end.y - start.y) / 2);
+
+        const elbow1 = Vec2.add(start, elbowDist);
+        const elbow2 = Vec2.sub(end, elbowDist);
+
+        path.moveTo(start.x, start.y);
+        path.lineTo(elbow1.x, elbow1.y);
+        path.lineTo(elbow2.x, elbow2.y);
+        path.lineTo(end.x, end.y);
     }
 }
